@@ -1,54 +1,32 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFrameworks>net5.0;net6.0</TargetFrameworks>
+    <TargetFrameworks>net6.0</TargetFrameworks>
     <IsPackable>false</IsPackable>
   </PropertyGroup>
   <PropertyGroup>
     <NoWarn>1701;1702;NU1608</NoWarn>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Logging" Version="5.0.0" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.11.0" />
-    <PackageReference Include="MSTest.TestAdapter" Version="2.2.7" />
-    <PackageReference Include="MSTest.TestFramework" Version="2.2.7" />
-    <PackageReference Include="coverlet.collector" Version="3.1.0">
+    <PackageReference Include="Microsoft.Extensions.Logging" Version="6.0.0" />
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.1.0" />
+    <PackageReference Include="MSTest.TestAdapter" Version="2.2.8" />
+    <PackageReference Include="MSTest.TestFramework" Version="2.2.8" />
+    <PackageReference Include="coverlet.collector" Version="3.1.2">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageReference>
-    <PackageReference Include="System.CommandLine" Version="2.0.0-beta1.21308.1" />
+    <PackageReference Include="System.CommandLine" Version="2.0.0-beta3.22106.2" />
   </ItemGroup>
   <ItemGroup>
     <!--<ProjectReference Include="..\..\..\command-line-api\src\System.CommandLine\System.CommandLine.csproj" />-->
     <ProjectReference Include="..\SqlBuildManager.Console\sbm.csproj" />
-    <ProjectReference Include="..\SqlSync.SqlBuild\SqlSync.SqlBuild.csproj" />
+    <ProjectReference Include="..\\SqlSync.SqlBuild\SQLSync.SqlBuild.csproj" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="..\TestConfig\**\*.*">
       <Link>TestConfig\%(RecursiveDir)%(FileName)%(Extension)</Link>
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </Content>
-  </ItemGroup>
-  <ItemGroup>
-    <Content Remove="..\TestConfig\sbm-linux-11.4.0.zip" />
-    <Content Remove="..\TestConfig\sbm-linux-12.1.0.zip" />
-    <Content Remove="..\TestConfig\sbm-linux-12.2.0.zip" />
-    <Content Remove="..\TestConfig\sbm-linux-13.0.0.zip" />
-    <Content Remove="..\TestConfig\sbm-linux-13.0.1.zip" />
-    <Content Remove="..\TestConfig\sbm-linux-13.0.2.zip" />
-    <Content Remove="..\TestConfig\sbm-linux-13.0.3.zip" />
-    <Content Remove="..\TestConfig\sbm-linux-13.1.0.zip" />
-    <Content Remove="..\TestConfig\sbm-linux-13.2.0.zip" />
-    <Content Remove="..\TestConfig\sbm-windows-11.4.0.zip" />
-    <Content Remove="..\TestConfig\sbm-windows-12.1.0.zip" />
-    <Content Remove="..\TestConfig\sbm-windows-12.2.0.zip" />
-    <Content Remove="..\TestConfig\sbm-windows-13.0.0.zip" />
-    <Content Remove="..\TestConfig\sbm-windows-13.0.1.zip" />
-    <Content Remove="..\TestConfig\sbm-windows-13.0.2.zip" />
-    <Content Remove="..\TestConfig\sbm-windows-13.0.3.zip" />
-    <Content Remove="..\TestConfig\sbm-windows-13.1.0.zip" />
-    <Content Remove="..\TestConfig\sbm-windows-13.2.0.zip" />
-    <Content Remove="..\TestConfig\secrests.yaml" />
-    <Content Remove="..\TestConfig\settings.txt" />
   </ItemGroup>
   <ItemGroup>
     <Compile Update="Properties\Resources.Designer.cs">
